Initial Home_Office 2
Changed the _site.yml by adding code_folding: hide, so that the code could be shown when necessary
