+++ _site.yml
@@ -14,4 +14,5 @@ navbar:
 output:
   html_document:
     theme: sandstone
+    code_folding: hide
 output_dir: "docs"
