+++ docs/data/desktop.ini
@@ -1,0 +1,5 @@
+[.ShellClassInfo]
+InfoTip=Dieser Ordner wird online freigegeben.
+IconFile=C:\Program Files\Google\Drive\googledrivesync.exe
+IconIndex=16
+
