--- docs/site_libs/dt-core-1.10.16/css/jquery.dataTables.extra.css
@@ -1,8 +1,0 @@
-/* Selected rows/cells */
-table.dataTable tr.selected td, table.dataTable td.selected {
-  background-color: #b0bed9 !important;
-}
-/* In case of scrollX/Y or FixedHeader */
-.dataTables_scrollBody .dataTables_sizing {
-  visibility: hidden;
-}
